+++ src/controllers/CsvImportController.php
@@ -140,6 +140,8 @@ class CsvImportController extends Controller
             }
         }
 
+        DatabaseTranslations::$plugin->databaseTranslationsService->afterUpdate();
+
         $this->setSuccessFlash("Imported $count new messages");
         return $this->redirect('database-translations');
     }