+++ src/controllers/TranslationsController.php
@@ -85,6 +85,8 @@ class TranslationsController extends Controller
             }
         }
 
+        DatabaseTranslations::$plugin->databaseTranslationsService->afterUpdate();
+
         if (count($errors)) {
             return $this->asJson([
                 'success' => false,
@@ -169,6 +171,7 @@ class TranslationsController extends Controller
     public function actionExportMigration()
     {
         $this->requirePostRequest();
+        $this->requirePermission('exportTranslationMigration');
         $filters = $this->request->post('filters');
         $languages = $this->request->post('languages', null);
         $sourceMessages = SourceMessage::filter($filters)->orderBy('message')->with('messages')->all();